+++ src/Interface.cs
@@ -24,4 +24,30 @@ namespace Dec
         /// </remarks>
         void PostCloneNew();
     }
+
+    /// <summary>
+    /// Used to name an object's entry in a Recorder's reference block.
+    /// </summary>
+    /// <remarks>
+    /// This applies only to objects that are actually written as references.
+    ///
+    /// Names should be unique within a single write, otherwise Dec will generate its own names.
+    /// </remarks>
+    public interface IRefName
+    {
+        /// <summary>
+        /// Returns the name for this object's reference entry, or null to accept a generated name.
+        /// </summary>
+        string RefName();
+    }
+
+    /// <summary>
+    /// Used to force an object into a Recorder's reference block even when only one reference to it exists.
+    /// </summary>
+    /// <remarks>
+    /// This does not set .Shared(); it has no effect if an object isn't .Shared().
+    /// </remarks>
+    public interface IRefForce
+    {
+    }
 }